requirements.txt
numpy
tensorflow
matplotlib
Pillow
streamlit
streamlit-drawable-canvas